+++ src/main/java/com/dynamics/domain/CategoriaDomain.java
@@ -1,0 +1,25 @@
+package com.dynamics.domain;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
+
+
+
+@Entity
+@Table(name="Categoria")
+public class CategoriaDomain extends PanacheEntityBase {
+
+    @Id
+    @GeneratedValue(strategy= GenerationType.IDENTITY)
+    @Column(name="Id")
+    public Short Id;
+
+    @Column(name="Descricao")
+    public String Descricao;
+}
