+++ src/main/java/com/dynamics/domain/TipoPromocaoDomain.java
@@ -1,17 +1,18 @@
-package com.dynamics.domain.categoria;
+package com.dynamics.domain;
 
 import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
 
 import javax.persistence.*;
 
 @Entity
-public class Categoria extends PanacheEntityBase {
+@Table(name = "TipoPromocao")
+public class TipoPromocaoDomain extends PanacheEntityBase {
 
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
     @Column(name="Id")
     public Short Id;
 
-    @Column(name="Descricao")
+    @Column(name = "Descricao")
     public String Descricao;
 }
