--- src/main/java/com/dynamics/domain/categoria/CategoriaResource.java
@@ -1,8 +1,0 @@
-package com.dynamics.domain.categoria;
-
-import com.dynamics.domain.categoria.Categoria;
-import io.quarkus.hibernate.orm.rest.data.panache.PanacheEntityResource;
-
-public interface CategoriaResource extends PanacheEntityResource<Categoria, Short> {
-
-}
