--- src/main/java/com/dynamics/domain/tipopromocao/TipoPromocao.java
@@ -1,18 +1,0 @@
-package com.dynamics.domain.tipopromocao;
-
-import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "TipoPromocao")
-public class TipoPromocao extends PanacheEntityBase {
-
-    @Id
-    @GeneratedValue(strategy= GenerationType.IDENTITY)
-    @Column(name="Id")
-    public Short Id;
-
-    @Column(name = "Descricao")
-    public String Descricao;
-}
