+++ src/main/java/com/dynamics/dto/ProdutoDTO.java
@@ -1,0 +1,7 @@
+package com.dynamics.dto;
+
+public class ProdutoDTO {
+
+	
+	
+}
