+++ src/main/java/com/dynamics/resource/CategoriaResource.java
@@ -1,0 +1,9 @@
+package com.dynamics.resource;
+
+import com.dynamics.domain.CategoriaDomain;
+
+import io.quarkus.hibernate.orm.rest.data.panache.PanacheEntityResource;
+
+public interface CategoriaResource extends PanacheEntityResource<CategoriaDomain, Short> {
+
+}
