+++ src/main/java/com/dynamics/resource/TipoPromocaoResource.java
@@ -1,17 +1,19 @@
-package com.dynamics.domain.tipopromocao;
+package com.dynamics.resource;
+
+import com.dynamics.domain.TipoPromocaoDomain;
 
 import io.quarkus.hibernate.orm.rest.data.panache.PanacheEntityResource;
 import io.quarkus.rest.data.panache.MethodProperties;
 import io.quarkus.rest.data.panache.ResourceProperties;
 
 @ResourceProperties(path = "tipopromocao")
-public interface TipoPromocaoResource extends PanacheEntityResource<TipoPromocao, Short> {
+public interface TipoPromocaoResource extends PanacheEntityResource<TipoPromocaoDomain, Short> {
 
     @MethodProperties(exposed = false)
-    TipoPromocao update(Short aShort, TipoPromocao tipoPromocao);
+    TipoPromocaoDomain update(Short aShort, TipoPromocaoDomain tipoPromocao);
 
     @MethodProperties(exposed = false)
-    TipoPromocao add(TipoPromocao tipoPromocao);
+    TipoPromocaoDomain add(TipoPromocaoDomain tipoPromocao);
 
     @MethodProperties(exposed = false)
     boolean delete(Short aShort);
